+++ app/src/androidTest/java/org/simple/clinic/patient/download/PatientLineListDownloaderTest.kt
@@ -11,7 +11,6 @@ import org.simple.clinic.TestClinicApp
 import org.simple.clinic.patient.download.PatientLineListDownloadResult.DownloadSuccessful
 import org.simple.clinic.rules.LocalAuthenticationRule
 import org.simple.sharedTestCode.util.Rules
-import java.time.LocalDate
 import javax.inject.Inject
 
 class PatientLineListDownloaderTest {
@@ -32,15 +31,9 @@ class PatientLineListDownloaderTest {
 
   @Test
   fun downloading_a_csv_should_work_correctly() {
-    // given
-    val bpCreatedAfter = LocalDate.parse("2018-01-01")
-    val bpCreatedBefore = LocalDate.parse("2018-03-01")
-
     // when
     val result = patientLineListDownloader
         .download(
-            bpCreatedAfter = bpCreatedAfter,
-            bpCreatedBefore = bpCreatedBefore,
             fileFormat = PatientLineListFileFormat.CSV
         )
         .blockingGet()
@@ -51,15 +44,9 @@ class PatientLineListDownloaderTest {
 
   @Test
   fun downloading_a_pdf_should_work_correctly() {
-    // given
-    val bpCreatedAfter = LocalDate.parse("2018-01-01")
-    val bpCreatedBefore = LocalDate.parse("2018-03-01")
-
     // when
     val result = patientLineListDownloader
         .download(
-            bpCreatedAfter = bpCreatedAfter,
-            bpCreatedBefore = bpCreatedBefore,
             fileFormat = PatientLineListFileFormat.PDF
         )
         .blockingGet()